Refactored file names and added CI workflow
